requirements.txt
brian2==2.7.1
numpy
scipy
matplotlib
pandas
pyyaml
pytest
